pyproject.toml
[project]
name = "machine-learning-for-process-engineering"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "supabase>=2.28.3",
    "torch>=2.11.0",
]
